Restore most of %__os_install_post
%__os_install_post was redefined to skip calling the parts that cause
strip to warn about unrecognized files. But this was done by copying and
modifying some very old definition of that macro. Instead, suppress the
brps that cause problems and let the rest of the macro stand.

In particular, this means that %__os_install_post_build_reproducibility
is retained and add-determinism is called, normalizing file ownership
and timestamps in static archives, fixing build reproducibility.